+++ modules/axruntime/src/stack.rs
@@ -41,6 +41,7 @@ impl KernelStack {
 }
 
 static mut IDLE_STACK: MaybeUninit<KernelStack> = MaybeUninit::uninit();
+static mut TASK_STACK: MaybeUninit<KernelStack> = MaybeUninit::uninit();
 
 pub fn init_idle_stack() -> Option<&'static KernelStack> {
     // Safety: single init during early boot.
@@ -48,5 +49,14 @@ pub fn init_idle_stack() -> Option<&'static KernelStack> {
         let stack = KernelStack::new()?;
         IDLE_STACK.write(stack);
         Some(IDLE_STACK.assume_init_ref())
+    }
+}
+
+pub fn init_task_stack() -> Option<&'static KernelStack> {
+    // Safety: single init during early boot.
+    unsafe {
+        let stack = KernelStack::new()?;
+        TASK_STACK.write(stack);
+        Some(TASK_STACK.assume_init_ref())
     }
 }